move evm chain into separate package
