+++ pkg/loop/internal/relayer/chains/evm/evm_chain.go
@@ -1,4 +1,4 @@
-package chains
+package evm
 
 import (
 	"context"
